refactor number config entity
